+++ packages/expo-router/src/link/preview/native.tsx
@@ -14,6 +14,7 @@ const areNativeViewsAvailable =
 export interface NativeLinkPreviewActionProps {
   identifier: string;
   title: string;
+  label?: string;
   icon?: string;
   image?: ImageRef | null;
   children?: React.ReactNode;